+++ packages/components/src/components/action-menu/action-menu.tsx
@@ -161,8 +161,7 @@ export class ActionMenu extends LitElement {
   @property() flipPlacements?: FlipPlacement[];
 
   /**
-   * Specifies an accessible label for the component.
-   *
+   * @copyDoc
    * @required
    */
   @property() label!: string;
@@ -200,9 +199,8 @@ export class ActionMenu extends LitElement {
    * Specifies the `calcite-action`s in the menu.
    *
    * @internal
-   * @readonly
    */
-  @property({ attribute: false }) get actions(): Action["el"][] {
+  @property() get actions(): Action["el"][] {
     return this._actions;
   }