+++ packages/components/src/components/navigation/navigation.tsx
@@ -82,7 +82,7 @@ export class Navigation extends LitElement {
 
   // #region Public Properties
 
-  /** When `navigationAction` is `true`, specifies an accessible label for the `calcite-action`. */
+  /** @copyDoc */
   @property() label?: string;
 
   /** When `true`, displays a `calcite-action` and emits a `calciteNavActionSelect` event on selection change. */