+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.0-next.50](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.49...@esri/calcite-components-react@5.2.0-next.50) (2026-08-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.49](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.48...@esri/calcite-components-react@5.2.0-next.49) (2026-08-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.48](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.47...@esri/calcite-components-react@5.2.0-next.48) (2026-08-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.47](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.46...@esri/calcite-components-react@5.2.0-next.47) (2026-08-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.46](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.45...@esri/calcite-components-react@5.2.0-next.46) (2026-08-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.45](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.44...@esri/calcite-components-react@5.2.0-next.45) (2026-08-11)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.44](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.43...@esri/calcite-components-react@5.2.0-next.44) (2026-08-11)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.43](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.42...@esri/calcite-components-react@5.2.0-next.43) (2026-08-06)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.42](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.41...@esri/calcite-components-react@5.2.0-next.42) (2026-08-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.41](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.40...@esri/calcite-components-react@5.2.0-next.41) (2026-08-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.40](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.39...@esri/calcite-components-react@5.2.0-next.40) (2026-07-31)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.39](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.38...@esri/calcite-components-react@5.2.0-next.39) (2026-07-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.38](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.37...@esri/calcite-components-react@5.2.0-next.38) (2026-07-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.37](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.36...@esri/calcite-components-react@5.2.0-next.37) (2026-07-30)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.2.0-next.36](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.35...@esri/calcite-components-react@5.2.0-next.36) (2026-07-28)
 
 **Note:** Version bump only for package @esri/calcite-components-react